chore: remove placeholder features
