+++ src/App.test.js
@@ -1,6 +1,6 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
+// import React from 'react'
+// import ReactDOM from 'react-dom'
+// import App from './App'
 
 it('renders without crashing', () => {
   /*
@@ -8,4 +8,4 @@ it('renders without crashing', () => {
   ReactDOM.render(<App />, div);
   ReactDOM.unmountComponentAtNode(div);
   */
-});
+})
