--- src/components/Browser/Whitelist.json
@@ -1,3 +1,0 @@
-[
-  "https://www.stateofthedapps.com"
-]
