--- src/components/Browser/index.js
@@ -1,3 +1,0 @@
-import Browser from './Browser'
-
-export default Browser
