--- src/components/Spinner.js
@@ -1,7 +1,0 @@
-import React, { Component } from 'react'
-
-export default class Spinner extends Component {
-  render() {
-    return <span>spinner</span>
-  }
-}
