--- src/containers/Dapps/DappItem.css
@@ -1,27 +1,0 @@
-.card {
-  display: grid;
-}
-.card-title {
-  font-size: 1.2rem;
-  padding: .5rem;
-  word-break: break-word;
-}
-.card-description {
-  padding: .5rem;
-  line-height: 1.6em;
-}
-.button {
-  display: block;
-  background-color: tomato;
-  padding: 10px 20px;
-  color: white;
-  text-decoration: none;
-  text-align: center;
-  transition: .3s ease-out;
-  height: 2.5rem;
-}
-
-.dapps .card img {
-  max-width: 100%;
-  margin: auto;
-}
