--- src/containers/Dapps/index.js
@@ -1,3 +1,0 @@
-import DappOverview from './DappOverview'
-
-export default DappOverview
