--- src/containers/Wallet/Account/CreateAccount.jsx
@@ -1,15 +1,0 @@
-import React, { Component } from 'react'
-import { FormCreateAccount } from 'ethereum-react-components'
-
-
-class CreateAccount extends Component {
-  render(){
-    return (
-    <main>
-      <FormCreateAccount />
-    </main>
-    )
-  }
-}
-
-export default CreateAccount
