--- src/containers/Wallet/AccountItem.css
@@ -1,70 +1,0 @@
-.account.card {
-  display: -ms-flexbox;
-  display: flex;
-  margin-left: 1rem;
-  margin-right: 1rem;
-  padding: .3rem;
-  background-color: white; 
-}
-
-.account h3 {
-  margin: 0px;
-}
-
-.account a {
-  color: black;
-}
-
-.account .account-identicon {
-  width: 2.3rem;
-  height: 2.3rem;
-  margin: auto;
-  padding: 30px;
-  text-align:center;
-  display: -ms-flexbox;
-  display: flex;
-  -ms-flex-direction: column;
-  flex-direction: column;
-  -ms-flex-wrap: nowrap;
-  flex-wrap: nowrap;
-  -ms-flex-pack: center;
-  justify-content: center;
-  -ms-flex-align: center;
-  align-items: center;
-  -ms-flex-line-pack: center;
-  align-content: center;
-}
-
-.account .account-information {
-  padding-left: .325rem;
-  overflow: hidden;
-  -ms-flex-positive: 1;
-  flex-grow: 1;
-}
-
-.account .account-information .account-address {
-  font-size: .8em;
-  line-height: 1.4em;
-  color: hsla(0,3%,49%,.6);
-}
-
-.account .account-information .account-balance {
-  font-size: 1rem;
-  line-height: 1.4em;
-  color: #2f2f2f
-}
-
-.account .identicon-medium {
-  background-clip: padding-box;
-  margin: 25px;
-}
-
-.account .identicon-medium img{
-  width: 32px;
-  height: 32px;
-}
-
-.account .account-action {
-  background: white;
-  margin-right: .5rem;
-}
