--- src/containers/Wallet/Wallet.css
@@ -1,11 +1,0 @@
-.wallet-actions .wallet-action{
-  background: white;
-  padding: .25rem;
-  margin-right: .25rem;
-  border: 1px solid #02a8f3;
-}
-
-.wallet-box-list{
-  overflow-y: scroll;
-  height: 70%;
-}
