--- src/containers/Wallet/index.js
@@ -1,3 +1,0 @@
-import Wallet from './Wallet'
-
-export default Wallet
